Banjax thread moved into AttackDetectin task
