+++ src/baskerville/models/pipeline_tasks/tasks.py
@@ -9,6 +9,7 @@ import datetime
 import itertools
 import json
 import os
+import threading
 import traceback
 
 import pyspark
@@ -21,7 +22,9 @@ import pyspark.sql.functions as psf
 
 from baskerville.db import get_jdbc_url
 from baskerville.db.models import RequestSet, Model
+from baskerville.models.banjax_report_consumer import BanjaxReportConsumer
 from baskerville.models.ip_cache import IPCache
+from baskerville.models.metrics.registry import metrics_registry
 from baskerville.models.pipeline_tasks.tasks_base import Task, MLTask, \
     CacheTask
 from baskerville.models.config import BaskervilleConfig
@@ -1253,12 +1256,66 @@ class AttackDetection(Task):
         self.df_chunks = []
         self.df_white_list = None
         self.ip_cache = IPCache(config, self.logger)
+        self.report_consumer = None
+        self.banjax_thread = None
+        self.register_metrics = config.engine.register_banjax_metrics
 
     def initialize(self):
         # super(SaveStats, self).initialize()
         if self.config.engine.white_list:
             self.df_white_list = self.spark.createDataFrame([[ip] for ip in set(self.config.engine.white_list)],
                                                             ['ip']).withColumn('white_list', F.lit(1))
+
+        self.report_consumer = BanjaxReportConsumer(self.config, self.logger)
+        if self.register_metrics:
+            self.register_banjax_metrics()
+        self.banjax_thread = threading.Thread(target=self.report_consumer.run)
+        self.banjax_thread.start()
+
+    def finish_up(self):
+        if self.banjax_thread:
+            self.banjax_thread.kill()
+            self.banjax_thread.join()
+
+        super().finish_up()
+
+    def register_banjax_metrics(self):
+        from baskerville.util.enums import MetricClassEnum
+
+        def incr_counter_for_ip_failed_challenge(metric, self, return_value):
+            metric.labels(return_value.get('value_ip'), return_value.get('value_site')).inc()
+            return return_value
+
+        consume_ip_failed_challenge_message = metrics_registry.register_action_hook(
+            self.report_consumer.consume_ip_failed_challenge_message,
+            incr_counter_for_ip_failed_challenge,
+            metric_name='ip_failed_challenge_on_website',
+            metric_cls=MetricClassEnum.counter,
+            labelnames=['ip', 'website']
+        )
+
+        setattr(self.report_consumer, 'consume_ip_failed_challenge_message', consume_ip_failed_challenge_message)
+
+        for field_name in self.report_consumer.status_message_fields:
+            target_method = getattr(self.report_consumer, f"consume_{field_name}")
+
+            def setter_for_field(field_name_inner):
+                def label_with_id_and_set(metric, self, return_value):
+                    metric.labels(return_value.get('id')).set(return_value.get(field_name_inner))
+                    return return_value
+
+                return label_with_id_and_set
+
+            patched_method = metrics_registry.register_action_hook(
+                target_method,
+                setter_for_field(field_name),
+                metric_name=field_name.replace('.', '_'),
+                metric_cls=MetricClassEnum.gauge,
+                labelnames=['banjax_id']
+            )
+
+            setattr(self.report_consumer, f"consume_{field_name}", patched_method)
+            self.logger.info(f"Registered metric for {field_name}")
 
     def classify_anomalies(self):
         self.logger.info('Anomaly thresholding...')
